64bit compile fixes with the new Teem version

+++ Core/Geom/NrrdTextureObj.cc
@@ -95,10 +95,10 @@ NrrdTextureObj::pad_to_power_of_2()
 {
   if (!nrrd_.get_rep() || !nrrd_->nrrd) return;
   NrrdDataHandle nout = scinew NrrdData();
-  int minp[3] = { 0, 0, 0 };
-  int maxp[3] = { 0, 
-                  Pow2(nrrd_->nrrd->axis[1].size)-1, 
-                  Pow2(nrrd_->nrrd->axis[2].size)-1 };
+  ptrdiff_t minp[3] = { 0, 0, 0 };
+  ptrdiff_t maxp[3] = { 0, 
+			Pow2(nrrd_->nrrd->axis[1].size)-1, 
+			Pow2(nrrd_->nrrd->axis[2].size)-1 };
 
   if (nrrdPad_nva(nout->nrrd, nrrd_->nrrd, minp, maxp, nrrdBoundaryBleed, 0)) {
     char *err = biffGetDone(NRRD);
